Safari-611.1.10/LayoutTests/fast/scrolling/home-end-keys-scroll-in-rtl-document-expected.txt

PASS: scrollLeft is -1208
